refactor offering catalog
