+++ source/OfferingCatalog.java
@@ -11,6 +11,17 @@ public class OfferingCatalog {
     public static List<Offering> getOfferings() {
         return offeringCatalog;
     }
+        public static List<Offering> getOfferings(Instructor instructor) {
+        List<Offering> offerings = new ArrayList<>();
+        for (Offering offering : OfferingCatalog.getOfferings()) {
+            for (City city : instructor.availabilities.cities) {
+                if ((offering.getLocation().getCity().getName()).equals(city.getName())) {
+                    offerings.add(offering);
+                }
+            }
+        }
+        return offerings;
+    }
 
     public static void removeOffering(Offering offering) {
         offeringCatalog.remove(offering);
